+++ packages/editor/editor-core/src/__tests__/unit/plugins/media/media-with-mock-facade.ts
@@ -1,3 +1,5 @@
+import { nextTick } from '@atlaskit/media-test-helpers';
+
 const removeOnCloseListener = jest.fn();
 const spies = {} as any;
 
@@ -77,9 +79,7 @@ describe('Media with mock facade', () => {
     const { pluginState } = editor(doc(p('{<>}')));
     await waitForAllPickersInitialised(pluginState);
 
-    const provider = await mediaProvider;
-    await provider.uploadContext;
-    await provider.viewContext;
+    await mediaProvider;
 
     expect(spies.popup.onClose).toHaveBeenCalledTimes(1);
     expect(spies.popup.onClose).toHaveBeenCalledWith(
@@ -93,9 +93,7 @@ describe('Media with mock facade', () => {
     const { pluginState } = editor(doc(p('{<>}')));
     await waitForAllPickersInitialised(pluginState);
 
-    const provider = await mediaProvider;
-    await provider.uploadContext;
-    await provider.viewContext;
+    await mediaProvider;
 
     pluginState.destroy();
     expect(removeOnCloseListener).toHaveBeenCalledTimes(1);
@@ -106,9 +104,8 @@ describe('Media with mock facade', () => {
     spies.dropzone.destroy.mockClear();
     const { pluginState } = editor(doc(p('{<>}')));
 
-    const provider = await mediaProvider;
-    await provider.uploadContext;
-    await provider.viewContext;
+    await mediaProvider;
+    await nextTick();
 
     pluginState.destroy();
     await Promise.all(pluginState.pickerPromises);
@@ -123,9 +120,7 @@ describe('Media with mock facade', () => {
     const { pluginState } = editor(doc(p('{<>}')));
     await waitForAllPickersInitialised(pluginState);
 
-    const provider = await mediaProvider;
-    await provider.uploadContext;
-    await provider.viewContext;
+    await mediaProvider;
 
     pluginState.showMediaPicker();
     expect(spies.popup.show).toHaveBeenCalledTimes(1);
@@ -139,9 +134,7 @@ describe('Media with mock facade', () => {
     const { pluginState } = editor(doc(p('{<>}')));
     await waitForAllPickersInitialised(pluginState);
 
-    const provider = await mediaProvider;
-    await provider.uploadContext;
-    await provider.viewContext;
+    await mediaProvider;
 
     pluginState.showMediaPicker();
     expect(spies.dropzone.activate).toHaveBeenCalledTimes(0);