+++ packages/editor/editor-core/src/plugins/media/ui/ResizableMediaSingle/types.ts
@@ -1,10 +1,10 @@
 import { MediaSingleLayout } from '@atlaskit/adf-schema';
 import { MediaSingleProps } from '@atlaskit/editor-common';
+import { MediaClientConfig } from '@atlaskit/media-core';
 import { EditorAppearance } from '../../../../types';
 import { GridType } from '../../../grid/types';
 import { EditorView } from 'prosemirror-view';
 import { EditorState } from 'prosemirror-state';
-import { Context } from '@atlaskit/media-core';
 
 export type EnabledHandles = { left?: boolean; right?: boolean };
 
@@ -23,6 +23,6 @@ export type Props = MediaSingleProps & {
   containerWidth: number;
   appearance?: EditorAppearance;
   selected: boolean;
-  viewContext?: Context;
+  viewMediaClientConfig?: MediaClientConfig;
   fullWidthMode?: boolean;
 };