+++ packages/editor/email-renderer/src/marks/text-color.ts
@@ -1,4 +1,5 @@
-import { createTag, serializeStyle } from '../util';
+import { createTag } from '../create-tag';
+import { serializeStyle } from '../serialize-style';
 import { MarkSerializerOpts } from '../interfaces';
 
 export default function textColor({ mark, text }: MarkSerializerOpts) {