+++ packages/editor/email-renderer/src/nodes/rule.ts
@@ -1,11 +1,14 @@
-import { colors } from '@atlaskit/theme';
-import { createTag, serializeStyle } from '../util';
+import { N30 } from '@atlaskit/adf-schema';
+import { createTag } from '../create-tag';
+import { createClassName } from '../styles/util';
 
-const css = serializeStyle({
-  border: 'none',
-  'border-bottom': `1px solid ${colors.N30A}`,
-});
+export const styles = `
+.${createClassName('rule')} {
+  border: none;
+  border-bottom: 1px solid ${N30};
+}
+`;
 
 export default function rule() {
-  return createTag('hr', { style: css });
+  return createTag('hr', { class: createClassName('rule') });
 }