+++ packages/editor/email-renderer/src/nodes/task-list.ts
@@ -1,5 +1,5 @@
 import { NodeSerializerOpts } from '../interfaces';
-import { createTag } from '../util';
+import { createTag } from '../create-tag';
 
 export default function taskList({ text }: NodeSerializerOpts) {
   return createTag('div', {}, text);