+++ packages/media/smart-card/src/view/CardWithUrl/component.tsx
@@ -1,5 +1,7 @@
 import * as React from 'react';
 import { MouseEvent, KeyboardEvent } from 'react';
+import LazilyRender from 'react-lazily-render';
+import { CardLinkView } from '@atlaskit/media-ui';
 
 import { CardWithUrlContentProps } from './types';
 import { uiCardClickedEvent } from '../../utils/analytics';
@@ -8,6 +10,25 @@ import { getDefinitionId, getServices } from '../../state/actions/helpers';
 import { BlockCard } from '../BlockCard';
 import { InlineCard } from '../InlineCard';
 import { useSmartLink } from '../../state';
+
+export function LazyCardWithUrlContent(props: CardWithUrlContentProps) {
+  const { appearance, isSelected, container, url } = props;
+  return (
+    <LazilyRender
+      offset={100}
+      component={appearance === 'inline' ? 'span' : 'div'}
+      placeholder={
+        <CardLinkView
+          isSelected={isSelected}
+          key={'lazy-render-placeholder'}
+          link={url}
+        />
+      }
+      scrollContainer={container}
+      content={<CardWithUrlContent {...props} />}
+    />
+  );
+}
 
 export function CardWithUrlContent({
   url,
@@ -33,7 +54,7 @@ export function CardWithUrlContent({
   };
   const handleClickWrapper = (event: MouseEvent | KeyboardEvent) => {
     handleAnalytics();
-    (onClick && onClick(event)) || handleClick(event);
+    onClick ? onClick(event) : handleClick(event);
   };
   const handleAuthorize = () => actions.authorize(appearance);
   // Lazily render into the viewport.