+++ packages/media/smart-card/src/view/__tests__/card-states.test.tsx
@@ -1,6 +1,4 @@
-jest.mock('react-lazily-render-scroll-parent', () => (data: any) =>
-  data.content,
-);
+jest.mock('react-lazily-render', () => (data: any) => data.content);
 jest.mock('react-transition-group/Transition', () => (data: any) =>
   data.children,
 );